+++ CHANGELOG.md
@@ -22,6 +22,7 @@ Changelog for NeoFS Node
 
 ### Removed
 - Drop creating new eacl tables with public keys (#3096)
+- BoltDB from write-cache (#3091)
 
 ### Updated
 
@@ -30,6 +31,12 @@ Using public keys as a rule target in eACL tables was deprecated, and
 since this realese it is not supported to create new eACL table with keys, 
 use addresses instead. 
 For more information call `neofs-cli acl extended create -h`.
+
+`small_object_size`, `workers_number`, `max_batch_size` and `max_batch_delay`
+paramteters are removed from `writecache` config. These parameters are related
+to the BoltDB part of the write-cache, which is dropped from the code. 
+Also, because of this, there will be automatic migration from BoltDB by flushing 
+objects to the main storage and removing database file.
 
 ## [0.44.2] - 2024-12-20